Now accepts OB generated test EIDAS certs
- Now has better regex based test for Authorization Number. This was
failing for non FR generated certs as the test was too FR specific.
- CertificateUtils now creates the qcType correctly, with the specific
qcType living in a sequence within the QCStatement that contains the
qcType.
- QCStatements.java now expects the specific qcType statment to be in a
DLSequence in the qcType's QCStatement.

NOTE: Currently the FR Directory generated EIDAS cert is not tested.
This is because I have yet to be able to generate a new cert with;
a) A correctly formatted Authorization Number.
b) The qcType formatted correctly.

This is because there is a somewhat chicken/egg situation. I would like
to generate that cert using the FR directory, but to do that I need this
lib to be re-issued and re-released.

A later commit will add that back into the test suite.